+++ src/testbench/NBLExpSpGEMM.cpp
@@ -359,7 +359,7 @@ void NBLExpSpGEMM::SpMM_DSParallel_RowStore_1P(const SpMatrix& A,
         int j    = A.ci[aij];
         int bjk  = B.rp[j];
         int bend = B.rp[j+1];
-        int vaij = A.v[aij];
+        double vaij = A.v[aij];
         for (; bjk < bend; ++bjk) {
           int k = B.ci[bjk];
           SPA.AddTo(k, vaij*B.v[bjk]);
